Auth/pyproject.toml
[tool.poetry]
name = "AuthService"
version = "0.1.0"
description = ""
authors = ["jungmir <ham4051@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.81.0"
hydra-core = "^1.2.0"
uvicorn = "^0.18.3"
rich = "^12.5.1"
httpx = "^0.23.0"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest-sugar = "^0.9.5"
pytest-watch = "^4.2.0"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
